Handle lists of contacts in file header.
Some people provide lists of authors / maintainers in their ASDF systems.
The generated reference file header now handles this properly (everyone is
listed as an author and a copyright owner).

The ASDF system mailto is DWIMly used as a contact address only when a
single author is found, and not associated with an email address explicitly.

2016-02-16  Didier Verna  <didier@didierverna.net>

	* core/src/declt.lisp (declt): Handle system maintainer and
	author slots as either strings or list of such. Pass lists of
	names and emails to RENDER-HEADER.
	(render-header): Update accordingly.