Protect against not finding component sources.
ASDF::COMPONENT-PATHNAME may return NIL when a component exists but we're
unable to locate it. Vendor-embedded versions of ASDF behave like that with
UIOP for instance. From now on, Declt is protected against this when
figuring out the sub-components of the system being documented: if a
dependency is not locatable, it's considered outside the system being
documented.

Since what we're actually documenting is supposed to be available from
source, I don't think we need to protect other calls to
COMPONENT-PATHNAME.
-- Spotted by Sabra Crolleton.

2016-02-22  Didier Verna  <didier@didierverna.net>

	* core/src/util/asdf.lisp (sub-component-p): Protect against
	ASDF::COMPONENT-PATHNAME returning NIL.