Improve the package collection scheme.
Previously, the system packages were collected by scanning for DEFPACKAGE
forms in the system(s) files. This breaks when packages are created in other
ways (such as wrappers around DEFPACKAGE).

The new scheme scans all live packages and collects those the source of
which belong to the system(s) being documented. This is much more robust and
removes the last bit of code walking from Declt.

2016-05-13  Didier Verna  <didier@didierverna.net>

	* core/src/declt.lisp (all-lisp-pathnames): New function.
	(add-packages): Use it.