Escape for Texinfo only on demand.
Previously, some strings were pre-escaped once and for all before use. That
was premature optimization at best, and it also hindered the readability of
the code. It's even better to /not/ do it for the future time where we may
output to something else than Texinfo.

This change of policy currently only affects declt.lisp. I haven't yet
checkded the other files.

2018-06-19  Didier Verna  <didier@didierverna.net>

	* core/src/declt.lisp (declt):
	(render-header): Don't pre-escape anything before use.