Handle heterogeneous generic functions in files definitions lists.
In a similar fashion to what the previous commit does for packages,
when computing the list of definitions belonging ti a file, we now
take into account the fact that generic functions and methods may
come from different files. As a result, a file definitions list now
includes not only generic functions, but also method definitions
directly.

2021-01-27  Didier Verna  <didier@didierverna.net>

	* core/src/extract/extract.lisp (finalize): [FILE-DEFINITION]
	Check for source file of methods as well as generic functions.