Fix generic function class upgrading.
Generic function definitions used to be wrongly upgraded to readers
or writers before all methods were actually added by the finalization
process. This led to cases where a generic function was seen as a reader
or a writer, only with regular methods, on which it was attempted to access
a non-existent target slot.

I fixed that by splitting the finalization process in two. Step one is now
called "stabilization". It's mostly the former finalization phase, less
the generic function class upgrading. Step two is now called "freezing",
and performs the aforementioned class upgrading, when we know for sure that
no new definition will be added.

2021-04-02  Didier Verna  <didier@didierverna.net>

	* extract/src/symbol.lisp
	(update-instance-for-different-class): Remove methods on method
	definitions. Don't upgrade the owner's class here.
	* extract/src/finalize.lisp (*stabilized*): Rename from *FINALIZED*
	(stabilize): Rename generic function and all methods from FINALIZE.
	(stabilize-clos-classoid-slot):
	(stabilize-clos-structure-slot): Rename functions from FINALIZE-*
	(freeze): New function.
	(finalize): New function. Stabilize and freeze definitions.
	* extract/src/extract.lisp (extract): Call FINALIZE.