Fix duplicate method combination documentation bug.
Method combinations... this has got to be the 42nd rewrite.
A generic function's method combination object do not uniquely represent
a method combination, but an instance of it with specific arguments.
As a consequence, multiple uses of the same combination, but with
different arguments, led to duplicate documentation. This patch fixes this
by identifying a method combination with its corresponding entry in SBCL's
SB-PCL::**METHOD-COMBINATIONS** hash table, instead of by one of its cache
object.

2021-04-05  Didier Verna  <didier@lrde.epita.fr>

	* extract/src/extract.lisp (make-symbol-definitions): Use SBCL's
	entry from the SB-PCL::**METHOD-COMBINATIONS** hash table to
	uniquely represent method combinations.
	* extract/src/symbol.lisp (make-combination-definition): Make sure
	we have at least one instantiation of the method combination.
	Deduce its type from there.
	(docstring): Retreive it from that instance.
	(identity-with-one-argument): Ditto.
	(lambda-list): Get it from the info structure directly.
	* extract/src/finalize.lisp (stabilize): Update generic functions
	and method combinations code for this new implementation.