Avoid potential duplication of system file definitions.
Some systems such as Arnesi explicitly list their system files as system
components. Since we already document them as pseudo-lisp files, we need to
filter them out. The drawback of that approach is that the system definition
becomes slightly incorrect, because we "hide" the appearance of the system
file as a component.

2021-04-05  Didier Verna  <didier@didierverna.net>

	* extract/src/extract.lisp (make-all-file-definitions): Filter out
	explicitly defined system file components.